chore: clean up fns
