+++ apps/web/src/state/farmsV4/state/accountPositions/type.ts
@@ -21,6 +21,9 @@ export type PositionDetail = {
   tokensOwed1: bigint
 
   // additional detail
+  /**
+   * @deprecated should rename to isFarmPool
+   */
   isStaked?: boolean
   chainId: number
   protocol: Protocol
